chore(suite): split tips components
